[refactor] shell erase command 파일명 변경
