+++ shell_core/commands/erase_command.py
