+++ tests/test_shell_erase.py
@@ -1,6 +1,6 @@
 import pytest
 from ssd_core.abstract_ssd import AbstractSSD
-from shell_core.commands.shell_erase import EraseCommand
+from shell_core.commands.erase_command import EraseCommand
 
 @pytest.mark.parametrize("lba,size", [(0,3),(1,8)])
 def test_erase_성공(mocker, lba, size):
@@ -15,7 +15,7 @@ def test_erase_성공(mocker, lba, size):
     assert ssd.erase.call_count == 1
 
 
-@pytest.mark.parametrize("lba,size,result", [(0, 11, 2),(8,44,5)])
+@pytest.mark.parametrize("lba,size,result", [(0, 11, 2),(8,44,5), (5, 1, 1)])
 def test_erase_10개초과_성공(mocker, lba, size, result):
     # arrange
     ssd: AbstractSSD = mocker.Mock()
@@ -28,7 +28,7 @@ def test_erase_10개초과_성공(mocker, lba, size, result):
     assert ssd.erase.call_count == result
 
 
-@pytest.mark.parametrize("lba,size,result", [(0, -11, 1),(8,-3,1), (22, -13, 2)])
+@pytest.mark.parametrize("lba,size,result", [(0, -11, 1),(8,-3,1), (22, -13, 2), (5, -1, 1)])
 def test_erase_size가음수인경우_성공(mocker, lba, size, result):
     # arrange
     ssd: AbstractSSD = mocker.Mock()
